requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_ipyw_jlab_nb_ext_conf=0.1.0=py36_0
alabaster=0.7.12=py36_0
anaconda-client=1.7.2=py36_0
anaconda-navigator=1.9.6=py36_0
anaconda-project=0.8.2=py36_0
asn1crypto=0.24.0=py36_0
astroid=2.1.0=py36_0
astropy=3.1.1=py36h7b6447c_0
atomicwrites=1.3.0=py_0
attrs=18.2.0=py36h28b3542_0
babel=2.6.0=py36_0
backcall=0.1.0=py36_0
backports=1.0=py36_1
backports.os=0.1.1=py36_0
backports.shutil_get_terminal_size=1.0.0=py36_2
beautifulsoup4=4.7.1=py36_1
bitarray=0.8.3=py36h14c3975_0
bkcharts=0.2=py36_0
blas=1.0=mkl
blaze=0.11.3=py36_0
bleach=3.1.0=py36_0
blosc=1.14.4=hdbcaa40_0
bokeh=1.0.4=py36_0
boost=1.59.0=py36_0
boto=2.49.0=py36_0
bottleneck=1.2.1=py36h035aef0_1
bzip2=1.0.6=h14c3975_5
ca-certificates=2019.1.23=0
cairo=1.14.8=0
certifi=2018.11.29=py36_0
cffi=1.11.5=py36he75722e_1
chardet=3.0.4=py36_1
click=7.0=py36_0
cloudpickle=0.7.0=py_0
clyent=1.2.2=py36_1
colorama=0.4.1=py36_0
conda=4.6.4=py36_0
conda-build=3.17.8=py36_0
conda-env=2.6.0=1
conda-verify=3.1.1=py36_0
contextlib2=0.5.5=py36_0
cryptography=2.3.1=py36hc365091_0
cudatoolkit=9.0=h13b8566_0
cudnn=7.3.1=cuda9.0_0
curl=7.61.0=h84994c4_0
cycler=0.10.0=pypi_0
cython=0.29.4=py36he6710b0_0
cytoolz=0.9.0.1=py36h14c3975_1
dask=1.1.1=py_0
dask-core=1.1.1=py_0
datashape=0.5.4=py36_1
dbus=1.10.20=0
decorator=4.3.2=py36_0
defusedxml=0.5.0=py36_1
distributed=1.25.3=py36_0
dlib=19.4=py36_0
docutils=0.14=py36_0
entrypoints=0.3=py36_0
et_xmlfile=1.0.1=py36_0
expat=2.2.6=he6710b0_0
fastcache=1.0.2=py36h14c3975_2
filelock=3.0.10=py36_0
flask=1.0.2=py36_1
flask-cors=3.0.7=py36_0
fontconfig=2.12.1=3
freetype=2.5.5=2
fribidi=1.0.5=h7b6447c_0
future=0.17.1=py36_0
get_terminal_size=1.0.0=haa9412d_0
gevent=1.4.0=py36h7b6447c_0
glib=2.50.2=1
glob2=0.6=py36_1
gmp=6.1.2=h6c8ec71_1
gmpy2=2.0.8=py36h10f8cd9_2
graphite2=1.3.12=h23475e2_2
greenlet=0.4.15=py36h7b6447c_0
gst-plugins-base=1.8.0=0
gstreamer=1.8.0=0
h5py=2.8.0=py36h989c5e5_3
harfbuzz=0.9.39=2
hdf5=1.10.2=hba1933b_1
heapdict=1.0.0=py36_2
html5lib=1.0.1=py36_0
icu=54.1=0
idna=2.8=py36_0
imageio=2.4.1=py36_0
imagesize=1.1.0=py36_0
importlib_metadata=0.7=py36_0
intel-openmp=2019.1=144
ipykernel=5.1.0=py36h39e3cac_0
ipython=7.2.0=py36h39e3cac_0
ipython_genutils=0.2.0=py36_0
ipywidgets=7.4.2=py36_0
isort=4.3.4=py36_0
itsdangerous=1.1.0=py36_0
jbig=2.1=hdba287a_0
jdcal=1.4=py36_0
jedi=0.13.2=py36_0
jeepney=0.4=py36_0
jinja2=2.10=py36_0
jpeg=9b=h024ee3a_2
jsonschema=2.6.0=py36_0
jupyter=1.0.0=py36_7
jupyter_client=5.2.4=py36_0
jupyter_console=6.0.0=py36_0
jupyter_core=4.4.0=py36_0
jupyterlab=0.35.3=py36_0
jupyterlab_server=0.2.0=py36_0
keyring=17.1.1=py36_0
kiwisolver=1.0.1=pypi_0
krb5=1.14.2=h56d4222_4
lazy-object-proxy=1.3.1=py36h14c3975_2
libarchive=3.3.2=hbf4e9fd_1
libcurl=7.61.0=h1ad7b7a_0
libedit=3.1.20170329=h6b74fdf_2
libffi=3.2.1=hd88cf55_4
libgcc=7.2.0=h69d50b8_2
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.14=0
liblief=0.9.0=h7725739_1
libpng=1.6.35=hbc83047_0
libsodium=1.0.16=h1bed415_0
libssh2=1.8.0=h9cfc8f7_4
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.9=he85c1e1_2
libtool=2.4.6=h7b6447c_5
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.4=0
libxslt=1.1.29=0
llvmlite=0.27.0=py36hd408876_0
locket=0.2.0=py36_1
lxml=4.1.1=py36h4d89739_0
lz4-c=1.8.1.2=h14c3975_0
lzo=2.10=h49e0be7_2
markupsafe=1.1.0=py36h7b6447c_0
matplotlib=3.0.2=pypi_0
mccabe=0.6.1=py36_1
mistune=0.8.4=py36h7b6447c_0
mkl=2018.0.3=1
mkl-service=1.1.2=py36h90e4bf4_5
mkl_fft=1.0.6=py36h7dd41cf_0
mkl_random=1.0.1=py36h4414c95_1
more-itertools=5.0.0=py36_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.1=hdf1c602_3
mpmath=1.1.0=py36_0
msgpack-python=0.6.1=py36hfd86e86_1
multipledispatch=0.6.0=py36_0
mypy=0.670=pypi_0
mypy-extensions=0.4.1=pypi_0
navigator-updater=0.2.1=py36_0
nbconvert=5.4.0=py36_1
nbformat=4.4.0=py36_0
nccl=1.3.5=cuda9.0_0
ncurses=6.1=he6710b0_1
networkx=2.2=py36_1
ninja=1.8.2=py36h6bb024c_1
nltk=3.4=py36_1
nose=1.3.7=py36_2
notebook=5.7.4=py36_0
numba=0.42.0=py36h962f231_0
numexpr=2.6.8=py36hd89afb7_0
numpy=1.16.1=pypi_0
numpydoc=0.8.0=py36_0
odo=0.5.1=py36_0
olefile=0.46=py36_0
opencv-python=4.0.0.21=pypi_0
openpyxl=2.5.14=py_0
openssl=1.0.2p=h14c3975_0
packaging=19.0=py36_0
pandas=0.24.1=py36he6710b0_0
pandoc=1.19.2.1=hea2e7c5_1
pandocfilters=1.4.2=py36_1
pango=1.40.3=1
parso=0.3.2=py36_0
partd=0.3.9=py36_0
patchelf=0.9=he6710b0_3
path.py=11.5.0=py36_0
pathlib2=2.3.3=py36_0
patsy=0.5.1=py36_0
pcre=8.39=1
pep8=1.7.1=py36_0
pexpect=4.6.0=py36_0
pickleshare=0.7.5=py36_0
pillow=4.2.1=py36_0
pip=19.0.1=py36_0
pixman=0.34.0=hceecf20_3
pkginfo=1.5.0.1=py36_0
pluggy=0.8.1=py36_0
ply=3.11=py36_0
prometheus_client=0.5.0=py36_0
prompt_toolkit=2.0.8=py_0
psutil=5.5.0=py36h7b6447c_0
ptyprocess=0.6.0=py36_0
py=1.7.0=py36_0
py-lief=0.9.0=py36h7725739_1
pycodestyle=2.5.0=py36_0
pycosat=0.6.3=py36h14c3975_0
pycparser=2.19=py36_0
pycrypto=2.6.1=py36h14c3975_9
pycurl=7.43.0.2=py36hb7f436b_0
pyflakes=2.1.0=py36_0
pygments=2.3.1=py36_0
pylint=2.2.2=py36_0
pyodbc=4.0.25=py36he6710b0_0
pyopenssl=19.0.0=py36_0
pyparsing=2.3.1=py36_0
pyqt=5.6.0=py36h22d08a2_6
pysocks=1.6.8=py36_0
pytables=3.4.4=py36ha205bf6_0
pytest=4.2.0=py36_0
pytest-arraydiff=0.3=py36h39e3cac_0
pytest-astropy=0.5.0=py36_0
pytest-doctestplus=0.2.0=py36_0
pytest-openfiles=0.3.2=py36_0
pytest-remotedata=0.3.1=py36_0
python=3.6.2=0
python-dateutil=2.8.0=pypi_0
python-libarchive-c=2.8=py36_6
pytorch=0.4.1=py36ha74772b_0
pytz=2018.9=py36_0
pywavelets=1.0.1=py36hdd07704_0
pyyaml=3.13=py36h14c3975_0
pyzmq=17.1.2=py36h14c3975_0
qt=5.6.2=5
qtawesome=0.5.6=py_0
qtconsole=4.4.3=py36_0
qtpy=1.6.0=py_0
readline=6.2=2
requests=2.21.0=py36_0
rope=0.11.0=py36_0
ruamel_yaml=0.15.46=py36h14c3975_0
scikit-image=0.13.1=py36h14c3975_1
scikit-learn=0.20.1=py36h4989274_0
scipy=1.1.0=py36hd20e5f9_0
seaborn=0.9.0=py36_0
secretstorage=3.1.0=py36_0
send2trash=1.5.0=py36_0
setuptools=40.8.0=pypi_0
simplegeneric=0.8.1=py36_2
singledispatch=3.4.0.3=py36_0
sip=4.18.1=py36hf484d3e_2
six=1.12.0=pypi_0
snappy=1.1.7=hbae5bb6_3
snowballstemmer=1.2.1=py36_0
sortedcollections=1.1.2=py36_0
sortedcontainers=2.1.0=py36_0
soupsieve=1.7.1=py36_0
sphinx=1.8.4=py36_0
sphinxcontrib=1.0=py36_1
sphinxcontrib-websupport=1.1.0=py36_1
spyder=3.3.3=py36_0
spyder-kernels=0.4.2=py36_0
sqlalchemy=1.2.17=py36h7b6447c_0
sqlite=3.13.0=0
statsmodels=0.9.0=py36h035aef0_0
sympy=1.3=py36_0
tblib=1.3.2=py36_0
terminado=0.8.1=py36_1
testpath=0.4.2=py36_0
tk=8.5.18=0
toolz=0.9.0=py36_0
torchvision=0.2.1=py_2
tornado=5.1.1=py36h7b6447c_0
tqdm=4.29.1=py_0
traitlets=4.3.2=py36_0
typed-ast=1.3.1=pypi_0
unicodecsv=0.14.1=py36_0
unixodbc=2.3.7=h14c3975_0
urllib3=1.24.1=py36_0
wcwidth=0.1.7=py36_0
webencodings=0.5.1=py36_1
werkzeug=0.14.1=py36_0
wheel=0.32.3=py36_0
widgetsnbextension=3.4.2=py36_0
wrapt=1.11.1=py36h7b6447c_0
wurlitzer=1.0.2=py36_0
xlrd=1.2.0=py36_0
xlsxwriter=1.1.2=py36_0
xlwt=1.3.0=py36_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zeromq=4.2.5=hf484d3e_1
zict=0.1.3=py36_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
